# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Market APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/order": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Create new order",
                "description": "Create new order",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "products",
                                    "products.*.product_id"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "products": {
                                        "description": "Product details",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity",
                                                "batch_id",
                                                "storage_id",
                                                "price"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": "2"
                                                },
                                                "batch_id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "storage_id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "price": {
                                                    "type": "numeric",
                                                    "example": "1000.00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error on validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Incorrect data."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error in server."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/product": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products",
                "description": "Get list of all available products.",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "batch_id",
                        "in": "query",
                        "description": "Id of Batch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "product_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Ahmad TEA"
                                                    },
                                                    "tech_params": {
                                                        "type": "string",
                                                        "example": "Made in England ..."
                                                    },
                                                    "quantity": {
                                                        "type": "integer",
                                                        "example": "10"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This batch does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profit": {
            "get": {
                "tags": [
                    "Profit"
                ],
                "summary": "Get profit",
                "description": "Get profit of each batch",
                "operationId": "getProfit",
                "parameters": [
                    {
                        "name": "batch_id",
                        "in": "query",
                        "description": "Id of Batch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "batch_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "batch_price": {
                                            "type": "numeric",
                                            "example": "10000.50"
                                        },
                                        "sales": {
                                            "type": "numeric",
                                            "example": "12000.20"
                                        },
                                        "profit": {
                                            "type": "numeric",
                                            "example": "2000.00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This batch does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}